Phase I and II

+++ app/controllers/articles_controller.rb
@@ -1,22 +1,17 @@
 class ArticlesController < ApplicationController
-  allow_unauthenticated_access only: %i[ index show ]
+  allow_unauthenticated_access only: %i[ index show noarticlehandler not_found_method ]
+
   def index
     @articles = Article.all
   end
   
   def show
-   url = request.original_url
-   if url.include?("/articles")
-     @url = "/articles"
-   else
-     @url = "/users"
-   end
-   begin 
-     @article = Article.find(params[:id])
-     @user = @article.user_id
-   rescue
-     redirect_to errorhandler_path
-   end
+    begin 
+      @article = Article.find(params[:id])
+      @user = @article.user_id
+    rescue
+      redirect_to noarticlehandler_path
+    end
   end
   
   def new
@@ -25,6 +20,11 @@ class ArticlesController < ApplicationController
       @article = Article.new(user_id: @user&.id)
     rescue
       redirect_to errorhandler_path
+    end
+    user = User.find_by(id: session[:user_id])
+    current_user = User.find(params[:user_id])
+    if current_user.id != user.id
+      redirect_to new_session_path
     end
   end
   
@@ -38,7 +38,14 @@ class ArticlesController < ApplicationController
   end
 
   def edit 
+    current_user = User.find_by(id: session[:user_id])
     @article = Article.find(params[:id])
+    editing_user = @article.user_id
+    if current_user.id == editing_user
+      'Do nothing'
+    else
+      redirect_to new_session_path
+    end
   end
 
   def update
@@ -52,11 +59,23 @@ class ArticlesController < ApplicationController
   
   def destroy
     article = Article.find(params[:id])
-    user = User.find(article.user_id)
-    article.destroy
-    redirect_to user
+    account_owner = User.find(article.user_id)
+    delete_user = User.find_by(id: session[:user_id])
+    if account_owner.id != delete_user.id
+      redirect_to new_session_path
+    else
+      article.destroy
+      redirect_to user_path(account_owner)
+    end
+  end
+
+  def noarticlehandler
   end
   
+  def not_found_method
+    render file: Rails.public_path.join('404.html'), status: :not_found, layout: true
+  end
+
   private
     def article_params
       params.expect(article: [:title, :content, :user_id])
